Fix type ambiguity

+++ src/types/sbitstream.jl
@@ -47,7 +47,7 @@ Base.typemax(::Type{T}) where {T<:SBitstream} = one(T)
 
 Base.isless(x::SBitstream, y::SBitstream) = Base.isless(float(x), float(y))
 Base.:(==)(x::SBitstream, y::SBitstream) = (float(x) == float(y))
-Base.hash(x::SBitstream, h) = hash(float(x), h)
+Base.hash(x::SBitstream, h::UInt) = hash(float(x), h)
 
 Base.show(io::IO, s::SBitstream) = print(io, "SBitstream($(s.value))")
 Base.show(io::IO, ::MIME"text/plain", s::SBitstream{T}) where T =
